Get rid of the Astroid dependency.
Astroid seemed like a really nice option because it promises to be able
to serialize ASTs back to source code. While it is true that it can do
this, it's not the same source code as went in, so it's not great for
things like refactoring tool, where you really don't want to mess with
the existing code, other than making the changes the user requested.