Only allow renaming at the definition site.
This will make imports so much simpler:

we'll "only" need to look at what other modules import this module,
and not through all the transitive imports this module does.

plus it makes less sense to rename at a point of usage, and if I get to
implementing go to definition, it won't actually be much of a loss.